+++ finagle-memcached/src/main/scala/com/twitter/finagle/memcached/protocol/Command.scala
@@ -7,7 +7,7 @@ sealed abstract class Command
 
 abstract class StorageCommand(key: ChannelBuffer, flags: Int, expiry: Time, value: ChannelBuffer) extends Command
 abstract class NonStorageCommand                                                      extends Command
-abstract class ArithmeticCommand(key: ChannelBuffer, delta: Int)                      extends NonStorageCommand
+abstract class ArithmeticCommand(key: ChannelBuffer, delta: Long)                      extends NonStorageCommand
 abstract class RetrievalCommand(keys: Seq[ChannelBuffer])                             extends NonStorageCommand
 
 case class Set(key: ChannelBuffer, flags: Int, expiry: Time, value: ChannelBuffer)     extends StorageCommand(key, flags, expiry, value)
@@ -20,5 +20,5 @@ case class Get(keys: Seq[ChannelBuffer])                                        
 case class Gets(keys: Seq[ChannelBuffer])                                             extends RetrievalCommand(keys)
 
 case class Delete(key: ChannelBuffer)                                                 extends Command
-case class Incr(key: ChannelBuffer, value: Int)                                       extends ArithmeticCommand(key, value)
-case class Decr(key: ChannelBuffer, value: Int)                                       extends ArithmeticCommand(key, -value)
+case class Incr(key: ChannelBuffer, value: Long)                                       extends ArithmeticCommand(key, value)
+case class Decr(key: ChannelBuffer, value: Long)                                       extends ArithmeticCommand(key, -value)
